# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: 'Machine-readable inventory of Flask `/api/v1` routes. Human-readable request/response examples: `docs/api/REFERENCE.md`.
    Tag taxonomy: `docs/api/openapi-taxonomy.md`.'
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/forum/tags/popular:
    get:
      tags:
      - Forum
      summary: GET /forum/tags/popular
      operationId: get_api_v1_forum_tags_popular
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/forum/tags/{slug}:
    get:
      tags:
      - Forum
      summary: GET /forum/tags/{slug}
      operationId: get_api_v1_forum_tags_slug
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
  /api/v1/forum/tags/{tag_id}:
    delete:
      tags:
      - Forum
      summary: DELETE /forum/tags/{tag_id}
      operationId: delete_api_v1_forum_tags_tag_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []